completed 4.21

+++ part4/controllers/blogs.js
@@ -1,35 +1,65 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const config = require('../utils/config')
+
+
+const jwt = require('jsonwebtoken')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
   if (!body.title || !body.url) {
     response.status(400).end()
   }
 
-  else {
-    const blog = new Blog({
-      "title": body.title,
-      "author": body.author,
-      "url": body.url,
-      "likes": body.likes || 0
-    })
+  const blog = new Blog({
+    "title": body.title,
+    "author": body.author,
+    "url": body.url,
+    "likes": body.likes || 0,
+    "user": user._id
+
+  })
   
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
-  }
+  const savedBlog = await blog.save()
+  // save to users aswell
+  user.blogs = user.blogs.concat(savedBlog._id) // add note to notes array in user
+  await user.save()
+
+  response.status(201).json(savedBlog)
 
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === user.id.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+
+  // return response.status(401).json({ error: 'token invalid' })
+
 })
 
 // updating likes of a post
